Move directorySink and tarSink into their own files.
